Combine createElement and JSX modules
There's a ton of overlap between the createElement implementation and
the JSX implementation, so I combined them into a single module.

In the actual build output, the shared code between JSX and
createElement will get duplicated anyway, because react/jsx-runtime and
react (where createElement livs) are separate, flat build artifacts.

So this is more about code organization — with a few key exceptions,
the implementations of createElement and jsx are highly coupled.
